fix: adminService 삭제 -> exeriseService로 통합
