+++ project/domain/src/main/java/com/project/exercise/exceptions/ExerciseAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.project.exercise.exceptions;
+
+public class ExerciseAlreadyExistsException extends RuntimeException {
+    public ExerciseAlreadyExistsException() {
+        super("이미 존재하는 운동입니다.");
+    }
+}
