+++ project/domain/src/main/java/com/project/exercise/service/ExerciseService.java
@@ -1,5 +1,7 @@
 package com.project.exercise.service;
 
+import com.project.configs.S3Manager;
+import com.project.exercise.exceptions.ExerciseAlreadyExistsException;
 import com.project.exercise.exceptions.ExerciseNotExistsException;
 import com.project.exercise.model.dto.DetailExerciseDto;
 import com.project.exercise.model.dto.SimpleExerciseDto;
@@ -10,7 +12,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.lang.reflect.Field;
 import java.util.Arrays;
 import java.util.List;
@@ -20,16 +24,27 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class ExerciseService {
 
+    private final S3Manager s3Manager;
     private final ExerciseRepository exerciseRepository;
 
 
-    public List<String> getExerciseCategory(){
+    @Transactional
+    public DetailExerciseDto saveExercise(String name, ExerciseCategory category, MultipartFile video) throws IOException {
+        if (exerciseRepository.findByName(name))
+            throw new ExerciseAlreadyExistsException();
+
+        String url = s3Manager.uploadFile(video);
+        Exercise exercise = exerciseRepository.save(new Exercise(null, name, url, category));
+        return new DetailExerciseDto(exercise);
+    }
+
+    public List<String> getExerciseCategory() {
         return Arrays.stream(ExerciseCategory.class.getFields()).map(Field::getName).collect(Collectors.toList());
     }
 
     @Transactional(readOnly = true)
-    @Cacheable(cacheNames = "exercise_list",key = "#category")
-    public List<SimpleExerciseDto> getExercises(ExerciseCategory category){
+    @Cacheable(cacheNames = "exercise_list", key = "#category")
+    public List<SimpleExerciseDto> getExercises(ExerciseCategory category) {
         return exerciseRepository.findAllByCategory(category).stream().map(SimpleExerciseDto::new).collect(Collectors.toList());
     }
